👌 IMPROVE: refactor supabase related code into separate package
